changes after CR comments
